pyproject.toml
[tool.poetry]
name = "llms-playground"
version = "0.1.0"
description = "Showcase and testing of several use cases of LLMs"
authors = ["Camilo Ramirez <camilo.ramirez@onlyfy.com>"]
readme = "README.md"
packages = [{include = "llms_playground"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.23.1"
langchain = {extras = ["docarray"], version = "^0.0.205"}
gpt4all = "^0.3.4"
ipykernel = "^6.23.2"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
tiktoken = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
